+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## 1.2.1
 - Fix unable to add recently used named imports.
 - Fix duplicate dependencies in package.json.
+- Add reading type definitions from `typings` in package.json.
 
 ## 1.2.0
 - Add recently used named imports for JavaScript/TypeScript node modules.